Migrate to setuptools built-in long_description
`setuptools-markdown` is deprecated now:
- https://dustingram.com/articles/2018/03/16/markdown-descriptions-on-pypi/

`setuptools >= 38.6.0` now has support to include markdown.

`setuptools-markdown` relies on a deprecated function
`pypandoc.convert`. Newer versions of pypandoc are not supporting this
function anymore and this broke the build. A patch is pending but
`setuptools-markdown` has been deprecated:
- https://github.com/msabramo/setuptools-markdown/pull/21
- https://pypi.org/project/setuptools-markdown/

In addition, the environment variable `NO_README` has been introduced
to allow to build and install `docker-inside` without adding
the markdown description (which is only required for releases on
pypi).